Refactor parsing methods
