+++ src/util.ts
@@ -1,0 +1,10 @@
+import { Stream } from 'stream'
+
+export function streamToString (stream: Stream): Promise<string> {
+  const chunks: Buffer[] = []
+  return new Promise((resolve, reject) => {
+    stream.on('data', chunk => chunks.push(chunk))
+    stream.on('error', reject)
+    stream.on('end', () => resolve(Buffer.concat(chunks).toString('utf8')))
+  })
+}
